+++ crates/affinidi-messaging/CHANGELOG.md
@@ -8,6 +8,14 @@ we find little issues that only affect deployment.
 Missing versions on the changelog simply reflect minor deployment changes on our
 tooling.
 
+## 24th June 2025
+
+### Mediator (0.10.11)
+
+* **SECURITY FIX:** DID Authentication flow was not checking inner plaintext from
+field attribute
+  * Will now check that the from field exists and matches signing key
+
 ## 6th June 2025
 
 ### Mediator (0.10.10)